Refactor pool class into dispatcher and pool
