+++ mlserver/parallel/pool.py
@@ -1,28 +1,24 @@
 import asyncio
 
-from itertools import cycle
-from asyncio import Future
-from concurrent.futures import ThreadPoolExecutor
 from multiprocessing import Queue
 from functools import wraps
-from typing import Any, Dict, Coroutine, Callable
+from typing import Any, Coroutine, Callable, Dict
 
 from ..model import MLModel
 from ..types import InferenceRequest, InferenceResponse
 from ..settings import Settings, ModelSettings
-from ..utils import get_wrapped_method, generate_uuid, schedule_with_callback
-from ..errors import InferenceError
+from ..utils import get_wrapped_method
 
 from .errors import InvalidParallelMethod
 from .worker import Worker
-from .utils import END_OF_QUEUE, terminate_queue, cancel_task, configure_inference_pool
+from .utils import configure_inference_pool
 from .messages import (
-    InferenceRequestMessage,
     InferenceResponseMessage,
     ModelUpdateMessage,
     ModelUpdateType,
 )
 from .logging import logger
+from .dispatcher import Dispatcher
 
 
 PredictMethod = Callable[[InferenceRequest], Coroutine[Any, Any, InferenceResponse]]
@@ -44,110 +40,23 @@ class InferencePool:
     def __init__(self, settings: Settings):
         configure_inference_pool(settings)
 
-        self._workers = {}
+        self._workers: Dict[int, Worker] = {}
         self._settings = settings
-        self._responses: Queue[InferenceResponseMessage] = Queue()
-        self._async_responses: Dict[str, Future[InferenceResponse]] = {}
-        self._executor = ThreadPoolExecutor()
+        responses: Queue[InferenceResponseMessage] = Queue()
         for idx in range(self._settings.parallel_workers):
             # TODO: Set callback to restart worker if it goes down (would
             # `worker.join` help with that?)
-            worker = Worker(self._responses)
+            worker = Worker(responses)
             worker.start()
-            self._workers[worker.pid] = worker
+            self._workers[worker.pid] = worker  # type: ignore
 
-        self._workers_round_robin = cycle(self._workers.keys())
-
-        # Start processing responses
-        self._start_processing_responses()
-
-    def _start_processing_responses(self):
-        self._active = True
-        self._process_responses_task = schedule_with_callback(
-            self._process_responses(), self._process_responses_cb
-        )
-
-    def _process_responses_cb(self, process_responses):
-        try:
-            process_responses.result()
-        except asyncio.CancelledError:
-            # NOTE: The response loop was cancelled from the outside, so don't
-            # restart
-            return
-        except Exception:
-            logger.exception("Response processing loop crashed. Restarting the loop...")
-            # If process loop crashed, restart it
-            self._start_processing_responses()
-
-    async def _process_responses(self):
-        logger.debug("Starting response processing loop...")
-        loop = asyncio.get_event_loop()
-        while self._active:
-            response = await loop.run_in_executor(self._executor, self._responses.get)
-
-            # If the queue gets terminated, detect the "sentinel value" and
-            # stop reading
-            if response is END_OF_QUEUE:
-                return
-
-            await self._process_response(response)
-
-    async def _process_response(self, response: InferenceResponseMessage):
-        internal_id = response.id
-
-        async_response = self._async_responses[internal_id]
-
-        if response.inference_response:
-            async_response.set_result(response.inference_response)
-        elif response.exception:
-            async_response.set_exception(response.exception)
-        else:
-            exc = InferenceError("Inference returned no value")
-            async_response.set_exception(exc)
+        self._dispatcher = Dispatcher(self._workers, responses)
+        self._dispatcher.start()
 
     async def predict(
         self, model_settings: ModelSettings, inference_request: InferenceRequest
     ) -> InferenceResponse:
-        internal_id = generate_uuid()
-
-        model_version = None
-        if model_settings.parameters:
-            model_version = model_settings.parameters.version
-
-        request_message = InferenceRequestMessage(
-            id=internal_id,
-            model_name=model_settings.name,
-            model_version=model_version,
-            inference_request=inference_request,
-        )
-
-        worker = self._get_worker()
-        worker.send_request(request_message)
-
-        loop = asyncio.get_running_loop()
-        async_response = loop.create_future()
-        self._async_responses[internal_id] = async_response
-
-        return await self._wait_response(internal_id)
-
-    def _get_worker(self) -> Worker:
-        """
-        Get next available worker.
-        By default, this is just a round-robin through all the workers.
-        """
-        worker_pid = next(self._workers_round_robin)
-        return self._workers[worker_pid]
-
-    async def _wait_response(self, internal_id: str) -> InferenceResponse:
-        async_response = self._async_responses[internal_id]
-
-        try:
-            inference_response = await async_response
-            return inference_response
-        finally:
-            del self._async_responses[internal_id]
-
-        return await async_response
+        return await self._dispatcher.predict(model_settings, inference_request)
 
     def parallel(self, f: PredictMethod):
         """
@@ -233,14 +142,8 @@ class InferencePool:
     async def close(self):
         logger.info("Waiting for inference pool shutdown")
         await self._close_workers()
-        await self._close_responses()
+        await self._dispatcher.stop()
         logger.info("Inference pool shutdown complete")
-
-    async def _close_responses(self):
-        await terminate_queue(self._responses)
-        await cancel_task(self._process_responses_task)
-        self._responses.close()
-        self._executor.shutdown()
 
     async def _close_workers(self):
         # First close down model updates loop
